Created singleton for global variables & constants
Tied up the structure and dependencies of the single scripts. A singleton script `tetrout.gd` is loaded which contains all useful constant and loads the basic block texture. Also created a "class" for Block objects that can be rendered just like normal Node2D's. This is useful for the "shoot" animation.