Huge update! :open_mouth:
Added touch screen controls that are visible when executing the game on a mobile device. I also managed to make it run on my Android phone! Other minor fixes were resizing the `single_block.png` image to 12x12 pixels and changing the way the texture is loaded in the script, because on Android it failed to display the texture.