Refactored TetrisCanvas
I have rewritten everything concerning the TetrisCanvas, to make it easier to understand and use for arcade mode. The blocks are no longer drawn by a Control node, instead we now use a TileMap. A single tile has it's own collision box, which makes it easier to detect collision between a block and the player. I also added a camera for smooth side-scrolling.